Move fake data to outer file
