+++ src/template/account/options_container/buying_table.jsx
@@ -15,172 +15,8 @@ import Typography from '@material-ui/core/Typography';
 import Checkbox from '@material-ui/core/Checkbox';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Switch from '@material-ui/core/Switch';
-import { nanoid } from 'nanoid';
 import { sort } from 'fast-sort';
-
-function randomDate(start, end)
-{
-  return new Date(start.getTime() + Math.random() * (end.getTime() - start.getTime()));
-}
-
-const headCells = [
-  {
-    id: 'productName',
-    numeric: false,
-    disablePadding: true,
-    label: 'Product name',
-  },
-  {
-    id: 'status',
-    numeric: false,
-    disablePadding: true,
-    label: 'Status',
-  },
-  {
-    id: 'purchaseDate',
-    numeric: true,
-    disablePadding: false,
-    label: 'Purchase date',
-  },
-  {
-    id: 'price',
-    numeric: true,
-    disablePadding: false,
-    label: 'Price ($)',
-  },
-];
-
-const rows = [
-  {
-    productId: nanoid(),
-    productName: 'Jordan 5 Retro Raging Bulls Red 2021 (GS)',
-    urkKey: 'air-jordan-5-retro-raging-bulls-red-2021-gs',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 140,
-    status: 'Shipping',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 5 Retro Raging Bulls Red 2021 (GS)',
-    urkKey: 'air-jordan-5-retro-raging-bulls-red-2021-gs',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 140,
-    status: 'Shipping',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 5 Retro Raging Bulls Red 2021 (GS)',
-    urkKey: 'air-jordan-5-retro-raging-bulls-red-2021-gs',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 140,
-    status: 'Shipping',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 5 Retro Raging Bulls Red 2021 (GS)',
-    urkKey: 'air-jordan-5-retro-raging-bulls-red-2021-gs',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 140,
-    status: 'Shipping',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 5 Retro Raging Bulls Red 2021 (GS)',
-    urkKey: 'air-jordan-5-retro-raging-bulls-red-2021-gs',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 140,
-    status: 'Shipping',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 5 Retro Raging Bulls Red 2021 (GS)',
-    urkKey: 'air-jordan-5-retro-raging-bulls-red-2021-gs',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 140,
-    status: 'Shipping',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 5 Retro Raging Bulls Red 2021 (GS)',
-    urkKey: 'air-jordan-5-retro-raging-bulls-red-2021-gs',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 140,
-    status: 'Shipping',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 5 Retro Raging Bulls Red 2021 (GS)',
-    urkKey: 'air-jordan-5-retro-raging-bulls-red-2021-gs',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 140,
-    status: 'Shipping',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 5 Retro Raging Bulls Red 2021 (GS)',
-    urkKey: 'air-jordan-5-retro-raging-bulls-red-2021-gs',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 140,
-    status: 'Shipping',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 5 Retro Raging Bulls Red 2021 (GS)',
-    urkKey: 'air-jordan-5-retro-raging-bulls-red-2021-gs',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 140,
-    status: 'Shipping',
-  },
-
-  {
-    productId: nanoid(),
-    productName: 'Jordan 9 Retro Change The World (W)\n',
-    urkKey: 'air-jordan-9-retro-change-the-world-w',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 190,
-    status: 'Received',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 9 Retro Change The World (W)\n',
-    urkKey: 'air-jordan-9-retro-change-the-world-w',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 190,
-    status: 'Received',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 9 Retro Change The World (W)\n',
-    urkKey: 'air-jordan-9-retro-change-the-world-w',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 190,
-    status: 'Received',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 9 Retro Change The World (W)\n',
-    urkKey: 'air-jordan-9-retro-change-the-world-w',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 190,
-    status: 'Received',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 9 Retro Change The World (W)\n',
-    urkKey: 'air-jordan-9-retro-change-the-world-w',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 190,
-    status: 'Received',
-  },
-  {
-    productId: nanoid(),
-    productName: 'Jordan 9 Retro Change The World (W)\n',
-    urkKey: 'air-jordan-9-retro-change-the-world-w',
-    purchaseDate: randomDate(new Date(2020, 0, 1), new Date()),
-    price: 190,
-    status: 'Received',
-  },
-];
+import { headCells, rows } from './fake_data';
 
 function stableSort(array, order, orderBy)
 {
